backend/requirements.txt
# ============================================
# IOB MAIIS - Backend Requirements
# Python 3.12 - Latest LTS Versions
# Updated: 2025-01-17
# ============================================

# ============================================
# WEB FRAMEWORK & SERVER
# ============================================
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.9
starlette==0.40.0

# ============================================
# DATABASE & ORM
# ============================================
sqlalchemy==2.0.35
asyncpg==0.29.0
alembic==1.13.3
psycopg2-binary==2.9.9
greenlet==3.1.1

# ============================================
# AUTHENTICATION & SECURITY
# ============================================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
pydantic[email]==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
cryptography==42.0.8

# ============================================
# REDIS CACHE & SESSION
# ============================================
redis==5.1.1
hiredis==3.0.0
aioredis==2.0.1

# ============================================
# VECTOR DATABASE
# ============================================
qdrant-client==1.12.0
grpcio==1.67.0
grpcio-tools==1.67.0

# ============================================
# AI/ML & LLM
# ============================================
langchain==0.3.7
langchain-community==0.3.7
langchain-core==0.3.15
sentence-transformers==3.2.1
transformers==4.46.2
torch==2.5.1
accelerate==1.1.1
safetensors==0.4.5
huggingface-hub==0.26.2

# ============================================
# EMBEDDINGS & NLP
# ============================================
openai==1.54.4
anthropic==0.39.0
cohere==5.11.2
tiktoken==0.8.0

# ============================================
# OCR & DOCUMENT PROCESSING
# ============================================
pytesseract==0.3.13
Pillow==11.0.0
pdf2image==1.17.0
pypdf==5.1.0
pypdfium2==4.30.0
python-docx==1.1.2
openpyxl==3.1.5
python-pptx==1.0.2
markdown==3.7

# ============================================
# SPEECH & AUDIO
# ============================================
openai-whisper==20231117
soundfile==0.12.1
librosa==0.10.2.post1
pydub==0.25.1
speechrecognition==3.10.4

# ============================================
# HTTP CLIENT & ASYNC
# ============================================
httpx==0.27.2
aiohttp==3.10.10
aiofiles==24.1.0
requests==2.32.3

# ============================================
# DATA VALIDATION & SERIALIZATION
# ============================================
pydantic==2.9.2
pydantic-core==2.23.4
email-validator==2.2.0
python-dateutil==2.9.0.post0

# ============================================
# UTILITIES & HELPERS
# ============================================
tenacity==9.0.0
retry==0.9.2
cachetools==5.5.0
python-slugify==8.0.4
python-magic==0.4.27

# ============================================
# MONITORING & LOGGING
# ============================================
prometheus-client==0.21.0
loguru==0.7.2
sentry-sdk[fastapi]==2.17.0
python-json-logger==3.1.0

# ============================================
# WEBSOCKETS & REAL-TIME
# ============================================
websockets==13.1
python-socketio==5.11.4
python-engineio==4.9.1

# ============================================
# BACKGROUND TASKS
# ============================================
celery==5.4.0
kombu==5.4.2
billiard==4.2.1

# ============================================
# TESTING
# ============================================
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
pytest-xdist==3.6.1
pytest-mock==3.14.0
pytest-timeout==2.3.1
faker==30.8.2
httpx==0.27.2
factory-boy==3.3.1

# ============================================
# CODE QUALITY & FORMATTING
# ============================================
black==24.10.0
flake8==7.1.1
isort==5.13.2
mypy==1.13.0
pylint==3.3.1
autopep8==2.3.1
bandit==1.7.10

# ============================================
# TYPE STUBS
# ============================================
types-redis==4.6.0.20241004
types-requests==2.32.0.20241016
types-python-dateutil==2.9.0.20241003
types-Pillow==10.2.0.20240822

# ============================================
# PERFORMANCE & OPTIMIZATION
# ============================================
orjson==3.10.10
ujson==5.10.0
msgpack==1.1.0

# ============================================
# RATE LIMITING
# ============================================
slowapi==0.1.9
limits==3.13.0

# ============================================
# CORS & SECURITY HEADERS
# ============================================
python-cors==1.0.0

# ============================================
# FILE HANDLING
# ============================================
python-magic-bin==0.4.14; sys_platform == 'win32'
chardet==5.2.0

# ============================================
# DATABASE MIGRATIONS
# ============================================
alembic==1.13.3

# ============================================
# ENVIRONMENT & CONFIG
# ============================================
python-decouple==3.8
environs==11.0.0

# ============================================
# PDF GENERATION (Optional)
# ============================================
reportlab==4.2.5
weasyprint==62.3

# ============================================
# EXCEL & CSV
# ============================================
pandas==2.2.3
xlsxwriter==3.2.0

# ============================================
# IMAGE PROCESSING
# ============================================
opencv-python-headless==4.10.0.84

# ============================================
# COMPRESSION
# ============================================
python-snappy==0.7.3

# ============================================
# SCHEDULING
# ============================================
apscheduler==3.10.4

# ============================================
# ADDITIONAL UTILITIES
# ============================================
arrow==1.3.0
pendulum==3.0.0
humanize==4.11.0
pytz==2024.2
